requirements.txt
APScheduler~=3.6.3
slack-webhook
python-dotenv
requests
